requirements.txt
streamlit>=1.29.0
google-cloud-aiplatform[adk,agent_engines]>=1.60.0
google-adk>=1.5.0
python-dotenv>=1.1.0
absl-py>=2.1.0
requests>=2.31.0
